requirements.txt
langchain_openai
langchain_core
langserve
python-dotenv
langchain_community
fastapi
streamlit
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu 
wikipedia
arxiv
